Fix victory theme according to types
